feat: change location for useCases for users
