+++ src/useCases/users/DeleteUser/DeleteUserUseCase.ts
@@ -1,16 +1,16 @@
 
-import User from '../../entity/User';
+import User from '../../../entity/User';
 import { getRepository } from 'typeorm';
-import { AppError } from '../../errors/AppError';
+import { AppError } from '../../../errors/AppError';
 
-export class GetUserUseCase {
+export class DeleteUserUseCase {
 
-    async execute(id: string): Promise<User> {
+    async execute(id: string): Promise<void> {
         const usersRepository = getRepository(User);
         
         const user: User = await usersRepository.findOne({ uuid: id });
         if (user) {
-            return user;
+            usersRepository.delete(user);
         } else {
             throw new AppError('User not found!');
         }
