+++ src/useCases/users/GetUser/GetUserUseCase.ts
@@ -1,16 +1,16 @@
 
-import User from '../../entity/User';
+import User from '../../../entity/User';
 import { getRepository } from 'typeorm';
-import { AppError } from '../../errors/AppError';
+import { AppError } from '../../../errors/AppError';
 
-export class DeleteUserUseCase {
+export class GetUserUseCase {
 
-    async execute(id: string): Promise<void> {
+    async execute(id: string): Promise<User> {
         const usersRepository = getRepository(User);
         
         const user: User = await usersRepository.findOne({ uuid: id });
         if (user) {
-            usersRepository.delete(user);
+            return user;
         } else {
             throw new AppError('User not found!');
         }
